testsp_perms_function/requirements.txt
functions-framework==3.*
google-api-python-client
google-cloud-storage
google-auth
google-cloud-bigquery
datetime
pytz
requests